refactor: extract PerformanceSeedData into its own file
Persistence.swift grew from 141 lines to 621 lines in this branch —
445 lines (72%) were the DEBUG-only seed harness. Moving the seed
types (PerformanceSeedConfiguration, PerformanceSeedData) into
PerformanceSeedData.swift reduces Persistence.swift back to its
pre-branch size (~173 lines) and keeps the harness self-contained.

No logic changes.
